# backend/src/openapi_server/models/blog_detail.py
# coding: utf-8

"""
    OSCHINA IT资讯APP后端API

    基于FastAPI开发的OSCHINA资讯代理后端API 1. 集成OSCHINA OAuth2.0授权登录 2. 后端代理所有OSCHINA OpenAPI请求，客户端无直接调用 3. 支持新闻/博客列表/详情、全局搜索 4. 自定义用户收藏功能（新闻/博客收藏管理） 5. 自定义数据模型适配Flutter客户端展示需求 

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, ConfigDict, Field, StrictInt, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List, Optional
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class BlogDetail(BaseModel):
    """
    博客详情数据
    """ # noqa: E501
    id: Optional[StrictInt] = None
    title: Optional[StrictStr] = None
    author: Optional[StrictStr] = None
    author_id: Optional[StrictInt] = Field(default=None, alias="authorId")
    publish_time: Optional[datetime] = Field(default=None, alias="publishTime")
    content: Optional[StrictStr] = Field(default=None, description="博客HTML内容")
    url: Optional[StrictStr] = Field(default=None, description="博客链接")
    favorite: Optional[StrictInt] = Field(default=None, description="0=未收藏 1=已收藏")
    comment_count: Optional[StrictInt] = Field(default=None, alias="commentCount")
    __properties: ClassVar[List[str]] = ["id", "title", "author", "authorId", "publishTime", "content", "url", "favorite", "commentCount"]

    @field_validator('favorite')
    def favorite_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value

        if value not in (0, 1,):
            raise ValueError("must be one of enum values (0, 1)")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of BlogDetail from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of BlogDetail from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "title": obj.get("title"),
            "author": obj.get("author"),
            "authorId": obj.get("authorId"),
            "publishTime": obj.get("publishTime"),
            "content": obj.get("content"),
            "url": obj.get("url"),
            "favorite": obj.get("favorite"),
            "commentCount": obj.get("commentCount")
        })
        return _obj