# backend/src/openapi_server/models/collect_request.py
# coding: utf-8

"""
    OSCHINA IT资讯APP后端API

    基于FastAPI开发的OSCHINA资讯代理后端API 1. 集成OSCHINA OAuth2.0授权登录 2. 后端代理所有OSCHINA OpenAPI请求，客户端无直接调用 3. 支持新闻/博客列表/详情、全局搜索 4. 自定义用户收藏功能（新闻/博客收藏管理） 5. 自定义数据模型适配Flutter客户端展示需求 

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictInt, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class CollectRequest(BaseModel):
    """
    收藏/取消收藏请求参数
    """ # noqa: E501
    target_type: StrictStr = Field(description="收藏目标类型", alias="targetType")
    target_id: StrictInt = Field(description="收藏目标ID", alias="targetId")
    __properties: ClassVar[List[str]] = ["targetType", "targetId"]

    @field_validator('target_type')
    def target_type_validate_enum(cls, value):
        """Validates the enum"""
        if value not in ('news', 'blog',):
            raise ValueError("must be one of enum values ('news', 'blog')")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of CollectRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of CollectRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "targetType": obj.get("targetType"),
            "targetId": obj.get("targetId")
        })
        return _obj