# news_check_app/api.yaml
openapi: 3.0.0
info:
  title: OSCHINA IT资讯APP后端API
  description: |
    基于FastAPI开发的OSCHINA资讯代理后端API
    1. 集成OSCHINA OAuth2.0授权登录
    2. 后端代理所有OSCHINA OpenAPI请求，客户端无直接调用
    3. 支持新闻/博客列表/详情、全局搜索
    4. 自定义用户收藏功能（新闻/博客收藏管理）
    5. 自定义数据模型适配Flutter客户端展示需求
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: https://api.your-domain.com
    description: 生产环境服务器

components:
  # 安全认证方案
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 后端自定义JWT认证（客户端登录后获取）
    OSCHINA_OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://www.oschina.net/action/oauth2/authorize
          tokenUrl: https://www.oschina.net/action/openapi/token
          scopes:
            read:news: 读取新闻资讯
            read:blog: 读取博客信息
            read:search: 搜索功能

  # 数据模型定义（自定义，适配Flutter客户端展示）
  schemas:
    # 通用响应体
    ApiResponse:
      type: object
      description: 全局统一API响应格式
      properties:
        code:
          type: integer
          example: 200
          description: 响应码 200=成功 400=参数错误 401=未授权 500=服务器错误
        msg:
          type: string
          example: 操作成功
          description: 响应信息
        data:
          type: object
          nullable: true
          description: 响应数据

    # 分页通用模型
    PageResult:
      type: object
      description: 分页数据统一格式
      properties:
        total:
          type: integer
          example: 100
          description: 总条数
        page:
          type: integer
          example: 1
          description: 当前页码
        pageSize:
          type: integer
          example: 20
          description: 每页条数
        list:
          type: array
          items:
            type: object
          description: 数据列表

    # 资讯列表通用项（新闻/博客共用，适配客户端展示：标题、缩略图、时间、阅读量等）
    InfoItem:
      type: object
      description: 新闻/博客列表展示项
      properties:
        id:
          type: integer
          example: 26754
          description: 内容ID
        title:
          type: string
          example: Flutter 最新版本特性解析
          description: 标题
        author:
          type: string
          example: 开源中国
          description: 作者
        publishTime:
          type: string
          format: date-time
          example: 2024-05-20 16:49:50
          description: 发布时间
        thumbnail:
          type: string
          example: https://www.oschina.net/default-thumbnail.png
          description: 缩略图(无则返回默认图)
        viewCount:
          type: integer
          example: 1200
          description: 阅读量
        commentCount:
          type: integer
          example: 25
          description: 评论数
        type:
          type: string
          enum: [news, blog]
          example: news
          description: 内容类型 news=新闻 blog=博客

    # 新闻详情
    NewsDetail:
      type: object
      description: 新闻详情数据
      properties:
        id:
          type: integer
          example: 26754
        title:
          type: string
          example: 新闻标题
        author:
          type: string
          example: 作者
        authorId:
          type: integer
          example: 253469
        publishTime:
          type: string
          format: date-time
        content:
          type: string
          description: 新闻HTML内容
        url:
          type: string
          description: 原文链接
        favorite:
          type: integer
          enum: [0,1]
          example: 0
          description: 0=未收藏 1=已收藏
        commentCount:
          type: integer
        relatedList:
          type: array
          items:
            $ref: '#/components/schemas/RelatedItem'

    # 博客详情
    BlogDetail:
      type: object
      description: 博客详情数据
      properties:
        id:
          type: integer
          example: 49090
        title:
          type: string
          example: 博客标题
        author:
          type: string
          example: 作者
        authorId:
          type: integer
          example: 100399
        publishTime:
          type: string
          format: date-time
        content:
          type: string
          description: 博客HTML内容
        url:
          type: string
          description: 博客链接
        favorite:
          type: integer
          enum: [0,1]
          example: 0
          description: 0=未收藏 1=已收藏
        commentCount:
          type: integer

    # 相关推荐项
    RelatedItem:
      type: object
      properties:
        title:
          type: string
        url:
          type: string

    # 搜索结果项
    SearchItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        publishTime:
          type: string
          format: date-time
        type:
          type: string
          enum: [news, blog, project, post]
        url:
          type: string

    # 收藏模型
    CollectRequest:
      type: object
      description: 收藏/取消收藏请求参数
      required:
        - targetType
        - targetId
      properties:
        targetType:
          type: string
          enum: [news, blog]
          example: news
          description: 收藏目标类型
        targetId:
          type: integer
          example: 26754
          description: 收藏目标ID

    # OAuth登录响应
    AuthToken:
      type: object
      properties:
        accessToken:
          type: string
          description: 后端JWT令牌
        expiresIn:
          type: integer
          example: 86400
          description: 过期时间(秒)
        userInfo:
          type: object
          description: 用户基础信息

  # 通用响应
  responses:
    Success:
      description: 请求成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: 未授权/登录失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 401
            msg: 请先登录
            data: null
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'

# API接口定义
paths:
  # ==================== 认证授权模块 ====================
  /auth/oschina/authorize-url:
    get:
      tags:
        - 认证授权
      summary: 获取OSCHINA授权链接
      description: 客户端调用此接口获取授权URL，通过WebView打开进行登录授权
      responses:
        200:
          $ref: '#/components/responses/Success'

  /auth/oschina/callback:
    get:
      tags:
        - 认证授权
      summary: OSCHINA授权回调接口
      description: 授权成功后回调，后端自动换取OSCHINA Token并生成客户端JWT
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: OSCHINA授权码
      responses:
        200:
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/AuthToken'

  /auth/logout:
    post:
      tags:
        - 认证授权
      summary: 退出登录
      security:
        - BearerAuth: []
      responses:
        200:
          $ref: '#/components/responses/Success'

  # ==================== 新闻模块 ====================
  /news/list:
    get:
      tags:
        - 新闻管理
      summary: 获取新闻列表
      description: 后端代理OSCHINA接口，返回自定义适配列表
      security:
        - BearerAuth: []
      parameters:
        - name: catalog
          in: query
          required: false
          schema:
            type: string
            default: 1
          description: 1=所有 2=综合新闻 3=软件更新
        - name: page
          in: query
          required: false
          schema:
            type: string
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: string
            default: 20
      responses:
        200:
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/PageResult'
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/InfoItem'
        401:
          $ref: '#/components/responses/Unauthorized'

  /news/detail/{id}:
    get:
      tags:
        - 新闻管理
      summary: 获取新闻详情
      description: 后端代理OSCHINA接口，返回自定义详情数据
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 新闻ID
      responses:
        200:
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/NewsDetail'
        401:
          $ref: '#/components/responses/Unauthorized'

  # ==================== 博客模块 ====================
  /blog/list:
    get:
      tags:
        - 博客管理
      summary: 获取博客列表
      description: 后端代理OSCHINA接口，返回自定义适配列表
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/PageResult'
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/InfoItem'
        401:
          $ref: '#/components/responses/Unauthorized'

  /blog/detail/{id}:
    get:
      tags:
        - 博客管理
      summary: 获取博客详情
      description: 后端代理OSCHINA接口，返回自定义详情数据
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 博客ID
      responses:
        200:
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/BlogDetail'
        401:
          $ref: '#/components/responses/Unauthorized'

  # ==================== 搜索模块 ====================
  /search:
    get:
      tags:
        - 搜索功能
      summary: 全局搜索
      description: 支持搜索新闻、博客，后端代理OSCHINA搜索接口
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: 搜索关键词
        - name: catalog
          in: query
          required: false
          schema:
            type: string
            default: news
          description: 搜索类型, 取值 news, blog, project, post
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: 搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/PageResult'
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/SearchItem'
        401:
          $ref: '#/components/responses/Unauthorized'

  # ==================== 收藏模块（自定义功能） ====================
  /collect/add:
    post:
      tags:
        - 收藏管理
      summary: 添加收藏（新闻/博客）
      description: 保存收藏类型+ID，后端关联用户
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectRequest'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthorized'

  /collect/remove:
    post:
      tags:
        - 收藏管理
      summary: 取消收藏
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectRequest'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthorized'

  /collect/list:
    get:
      tags:
        - 收藏管理
      summary: 获取我的收藏列表
      description: 后端根据收藏ID批量请求OSCHINA，返回列表数据
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/PageResult'
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/InfoItem'
        401:
          $ref: '#/components/responses/Unauthorized'